[Refact] Code Refactoring
