--- README.md
@@ -1,53 +1,0 @@
-# "반려동물과 떠나는 완벽한 여행 | PETTY" ⚡
-
-![Image](https://github.com/user-attachments/assets/cfbf9440-5a3e-4802-9363-f2abb7861159)
-
-----
-
-
-
-### 파일 정리
-
-1. 베이스 UI/UX
-
-   ```shell
-   common.css
-   teplates/
-   base-tem.html
-   ```
-
-2. 메인 페이지
-
-   ```shell
-   index.html
-   index.css
-   ```
-
-3. 여행지/숙소/음식점 추천
-
-   ```shell
-   reco.html
-   reco.js
-   reco.css
-   ```
-
-4. 이미지 분석
-   ```shell
-   vision.html
-   ```
-
-6. 커뮤니티
-
-   ```shell
-   community.html
-   community.js
-   community.css
-   ```
-
-7. 로그인 / 회원가입
-   ```shell
-   login.html
-   siginin.html
-   script.js
-   ```
-
