--- base-tem.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
-    <!-- favicon -->
-    <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
-    />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="./asset/favicon/apple-touch-icon.png"
-    />
-    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
-    <meta
-      property="og:title"
-      content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
-    />
-    <meta
-      property="og:description"
-      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 최적의 숙소, 관광지, 음식점을 추천해 줄게요! 🐾"
-    />
-    <meta
-      property="og:image"
-      content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
-    />
-    <link rel="stylesheet" href="./style.css" />
-
-    <!-- Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-    <!-- Bootstrap Icons CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
-    />
-
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
-
-    <link rel="stylesheet" href="style.css" />
-    <link rel="stylesheet" href="common.css" />
-  </head>
-  <!-- ====== body ======= -->
-  <body>
-    <!-- nav bar - js -->
-    <div id="navbar"></div>
-
-    <!-- nav bar - js -->
-    <div id="footer"></div>
-    <script src="./script.js" defer></script>
-    <script src="./main.js"></script>
-    <!-- Bootstrap JS (Optional, for features like dropdowns) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  </body>
-</html>
