+++ community.html
@@ -1,91 +1,93 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
-    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"></script>
-    <!-- favicon -->
+
+    <!-- ✅ Supabase SDK -->
+    <script
+      defer
+      src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"
+    ></script>
+
+    <!-- ✅ Custom JS (모듈 분리 가정) -->
+    <script type="module" src="./js/main.js" defer></script>
+    <script type="module" src="./js/post.js" defer></script>
+    <script defer src="./js/script.js"></script>
+
+    <!-- ✅ Bootstrap -->
     <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
     />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
+    />
+    <script
+      defer
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
+    ></script>
+
+    <!-- ✅ Favicon -->
+    <link rel="icon" href="./asset/favicon/favicon.ico" />
     <link
       rel="apple-touch-icon"
       sizes="180x180"
       href="./asset/favicon/apple-touch-icon.png"
     />
     <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
+
+    <!-- ✅ OG Meta Tags -->
     <meta
       property="og:title"
       content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
     />
     <meta
       property="og:description"
-      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 최적의 숙소, 관광지, 음식점을 추천해 줄게요! 🐾"
+      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 도와줄게요!"
     />
     <meta
       property="og:image"
       content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
     />
-    <!-- <link rel="stylesheet" href="./style.css" /> -->
 
-    <!-- Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-    <!-- Bootstrap Icons CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
-    />
+    <!-- ✅ Styles -->
+    <link rel="stylesheet" href="./common.css" />
+    <link rel="stylesheet" href="./community.css" />
 
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
-    <link rel="stylesheet" href="common.css" />
-    <link rel="stylesheet" href="community.css" />
+    <title>PETTY | 집사들의 이야기</title>
   </head>
-  <!-- ====== body ======= -->
+
   <body>
-    <!-- nav bar - js -->
+    <!-- ✅ 네비게이션 -->
     <div id="navbar"></div>
 
+    <!-- ✅ 상단 배너 -->
     <section class="container mt-4">
-      <img src="./asset/006.png" class="img-fluid" />
+      <img src="./asset/006.png" class="img-fluid" alt="커뮤니티 배너" />
     </section>
+
+    <!-- ✅ 게시글 추가 영역 -->
     <section class="container form-container">
       <h4>🐈 집사들의 글</h4>
       <button id="addPostBtn" class="btn btn-primary">게시글 추가</button>
     </section>
 
-    <div id="mainContainer">
-      <div
-        id="spinnerContainer"
-        class="text-center mt-3"
-        style="display: block"
-      >
+    <!-- ✅ 게시글 목록 -->
+    <main id="mainContainer">
+      <div id="spinnerContainer" class="text-center mt-3">
         <div class="spinner-border text-success" role="status">
           <span class="visually-hidden">Loading...</span>
         </div>
       </div>
+
       <div class="container">
         <div class="row" id="postList"></div>
-        <!-- 🔹 한 줄에 3개씩 배치 -->
       </div>
-    </div>
+    </main>
 
-    <!-- nav bar - js -->
+    <!-- ✅ 푸터 -->
     <div id="footer"></div>
-    <!-- <script src="./community.js" defer></script> -->
-    <script src="./main.js"></script>
-    <script src="./post.js"></script>
-
-    <!-- Bootstrap JS (Optional, for features like dropdowns) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
   </body>
 </html>
