+++ css/common.css
@@ -1,0 +1,101 @@
+@font-face {
+  font-family: "HakgyoansimDunggeunmisoTTF-B";
+  src: url("https://fastly.jsdelivr.net/gh/projectnoonnu/2408-5@1.0/HakgyoansimDunggeunmisoTTF-B.woff2")
+    format("woff2");
+  font-weight: 700;
+  font-style: normal;
+}
+@font-face {
+  font-family: "Pretendard-Regular";
+  src: url("https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff")
+    format("woff");
+  font-weight: 400;
+  font-style: normal;
+}
+@font-face {
+  font-family: "GmarketSansMedium";
+  src: url("https://fastly.jsdelivr.net/gh/projectnoonnu/noonfonts_2001@1.1/GmarketSansMedium.woff")
+    format("woff");
+  font-weight: normal;
+  font-style: normal;
+}
+body {
+  font-family: "Pretendard-Regular";
+  background-color: #fcf2d74d !important;
+  min-height: 100vh;
+  margin: 0;
+  display: flex;
+  flex-direction: column;
+  word-break: keep-all;
+}
+
+.main {
+  flex: 1;
+}
+h2 {
+  font-family: "HakgyoansimDunggeunmisoTTF-B";
+}
+
+a {
+  text-decoration-line: none;
+}
+
+/* Navigation Bar */
+.navbar {
+  background-color: #ffe18e !important;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  font-size: 18px;
+}
+
+.nav-logo-img {
+  max-height: 50px;
+}
+
+.navbar-brand {
+  font-weight: bold;
+  color: #1e3a8a;
+}
+
+.nav-link {
+  color: #333;
+  transition: color 0.3s ease;
+}
+
+.nav-link:hover {
+  color: #fbc63c;
+}
+
+.nav .btn {
+  padding: 0px;
+  margin-bottom: 0px;
+}
+
+.btn-dark {
+  background-color: #fbc63c;
+  border-color: #fbc63c;
+}
+
+/* Footer */
+.footer {
+  background-color: #ffe18e;
+  padding: 30px 0;
+  text-align: center;
+  color: #666;
+  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
+}
+
+/* Responsive Design */
+@media (max-width: 768px) {
+  .input-group {
+    width: 80% !important;
+  }
+  .carousel img {
+    height: 30% !important;
+  }
+}
+
+@media (max-width: 576px) {
+  .carousel img {
+    height: 250px;
+  }
+}
