+++ css/index.css
@@ -1,0 +1,140 @@
+/* 캐러셀 컨테이너 */
+.carousel-container {
+  width: 90%;
+  max-width: 1200px;
+  margin: 0 auto;
+  overflow: hidden;
+}
+
+/* 캐러셀 이미지 스타일 */
+.carousel img {
+  object-fit: cover;
+  height: 600px;
+}
+
+.container {
+  max-width: 1200px;
+}
+
+.col-md-4 {
+  border-radius: 15px;
+}
+.col-md-4 :hover {
+  background-color: rgba(255, 255, 230, 0.459);
+}
+/* Features Section */
+/* .features .card {
+      border: none;
+      border-radius: 15px;
+      box-shadow: 0 4px 8px #fbc63c;
+      transition: transform 0.3s ease, box-shadow 0.3s ease;
+      margin-bottom: 20px;
+    }
+    
+    .features .card:hover {
+      transform: translateY(-5px);
+      box-shadow: 0 6px 12px #fbc63c;
+    }
+    
+    .features h5 {
+      color: #000000;
+    } */
+
+/* content */
+
+.features {
+  min-height: 600px;
+}
+
+.features {
+  padding-top: 20px;
+  padding-bottom: 20px;
+}
+.features .card {
+  transition: transform 0.3s ease, box-shadow 0.3s ease;
+  margin-bottom: 20px;
+  height: 100%; /* 카드 높이 통일 */
+  display: flex;
+  flex-direction: column;
+  margin-top: 20px;
+  border: 0px;
+}
+
+.features .card:hover {
+  transform: translateY(-5px);
+}
+
+.features h5 {
+  color: #000000;
+}
+
+.features .card img {
+  width: 100%;
+  height: 250px; /* 원하는 이미지 높이 설정 */
+  object-fit: cover; /* 넘치는 이미지 비율 유지하면서 맞춤 */
+}
+
+.features .card-body {
+  flex: 1; /* 내용이 균등하게 배치되도록 설정 */
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  text-align: center;
+  padding: 10px;
+}
+
+.features .row {
+  gap: 30px; /* 카드 간격을 조정 */
+}
+
+.features .col-md-4 {
+  padding: 0 20px; /* 좌우 간격 추가 */
+}
+
+/* Success Stories Section */
+.success-stories img {
+  width: 100%;
+  height: 200px;
+  object-fit: cover;
+  border-radius: 10px;
+  margin-bottom: 1rem;
+}
+
+.success-stories h5 {
+  color: #1e3a8a;
+}
+
+/* Quick Search Section */
+.input-group {
+  border-radius: 25px;
+  overflow: hidden;
+}
+
+.card :hover {
+  background-color: rgba(255, 255, 230, 0.459);
+}
+
+@media (max-width: 768px) {
+  .col-md-4 {
+    width: 70%;
+  }
+  .features {
+    align-items: center;
+    justify-content: center;
+  }
+}
+
+@media (max-width: 576px) {
+  .carousel img {
+    height: 250px;
+  }
+}
+
+#navbarNav.collapse.navbar-collapse {
+  font-family: "HakgyoansimDunggeunmisoTTF-B";
+}
+
+#login-btn.btn.btn-success {
+  background-color: #a3520f;
+  border: none;
+}
