+++ css/login.css
@@ -1,0 +1,14 @@
+.login-container {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  align-items: center;
+  justify-content: center;
+}
+#navbarNav.collapse.navbar-collapse {
+  font-family: "HakgyoansimDunggeunmisoTTF-B";
+}
+#login-btn.btn.btn-success {
+  background-color: #a3520f;
+  border: none;
+}
