+++ index.html
@@ -1,26 +1,12 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <!-- ① 메타 태그 -->
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
-    <!-- favicon -->
-    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"></script>
-    <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
-    />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="./asset/favicon/apple-touch-icon.png"
-    />
-    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
+    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
+
+    <!-- ② OG 태그 (공유 썸네일용) -->
     <meta
       property="og:title"
       content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
@@ -33,20 +19,29 @@
       property="og:image"
       content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
     />
-    <!-- <link rel="stylesheet" href="./style.css" /> -->
 
-    <!-- Bootstrap CSS -->
+    <!-- ③ Favicon -->
+    <link
+      rel="icon"
+      type="image/png"
+      href="./asset/favicon/favicon-96x96.png"
+    />
+    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
+    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
+    <link rel="apple-touch-icon" href="./asset/favicon/apple-touch-icon.png" />
+    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
+
+    <!-- ④ 외부 라이브러리 (CSS) -->
     <link
       href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
       rel="stylesheet"
     />
-    <!-- Bootstrap Icons CSS -->
     <link
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
     />
 
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
+    <!-- ⑤ 프로젝트 CSS -->
     <link rel="stylesheet" href="common.css" />
     <link rel="stylesheet" href="index.css" />
   </head>
@@ -152,7 +147,7 @@
         </div>
 
         <div class="col-md-4 col-12 mb-4">
-          <a href="community.html">
+          <a href="./community.html">
             <div class="card">
               <img src="./asset/card3.jpg" class="img-fluid" alt="커뮤니티" />
               <div class="card-body">
@@ -167,10 +162,13 @@
 
     <!-- nav bar - js -->
     <div id="footer"></div>
-    <script src="./script.js"></script>
-    <script src="./main.js"></script>
-    <script src="./post.js"></script>
-    <!-- Bootstrap JS (Optional, for features like dropdowns) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"></script>
+    <script src="./js/main.js" defer></script>
+    <script src="./js/post.js" defer></script>
+    <script src="./js/script.js" defer></script>
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
+      defer
+    ></script>
   </body>
 </html>
