+++ post-detail.html
@@ -1,54 +1,57 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
-    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"></script>
-    <!-- favicon -->
+
+    <!-- ✅ Supabase -->
+    <script
+      defer
+      src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"
+    ></script>
+
+    <!-- ✅ 앱 스크립트 -->
+    <script defer type="module" src="./js/main.js"></script>
+    <script defer type="module" src="./js/post.js"></script>
+    <script defer type="module" src="./js/post-detail.js"></script>
+    <script defer src="./js/script.js"></script>
+
+    <!-- ✅ Bootstrap CSS -->
     <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
     />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
+    />
+
+    <!-- ✅ Favicon -->
+    <link rel="icon" href="./asset/favicon/favicon.ico" />
     <link
       rel="apple-touch-icon"
       sizes="180x180"
       href="./asset/favicon/apple-touch-icon.png"
     />
     <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
+
+    <!-- ✅ OG 태그 -->
     <meta
       property="og:title"
       content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
     />
     <meta
       property="og:description"
-      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 최적의 숙소, 관광지, 음식점을 추천해 줄게요! 🐾"
+      content="우리 아이와 여행을 떠나고 싶다면? PETTY가 도와줄게요 🐾"
     />
     <meta
       property="og:image"
       content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
     />
-    <link rel="stylesheet" href="./style.css" />
-    <link rel="stylesheet" href="./community.css" />
-    <!-- Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-    <!-- Bootstrap Icons CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
-    />
 
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
+    <!-- ✅ 스타일 -->
     <link rel="stylesheet" href="common.css" />
-    <link rel="stylesheet" href="community.css" />
+    <link rel="stylesheet" href="write.css" />
   </head>
   <!-- ====== body ======= -->
   <body>
@@ -89,10 +92,6 @@
 
     <!-- nav bar - js -->
     <div id="footer"></div>
-    <!-- <script src="./community.js" defer></script> -->
-    <script src="./main.js"></script>
-    <script src="./post.js"></script>
-    <script src="./post-detail.js"></script>
 
     <!-- Bootstrap JS (Optional, for features like dropdowns) -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
