+++ post.html
@@ -1,0 +1,132 @@
+<!DOCTYPE html>
+<html lang="ko">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <!-- ✅ Supabase -->
+    <script
+      defer
+      src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"
+    ></script>
+
+    <!-- ✅ 앱 스크립트 -->
+    <script defer type="module" src="./js/main.js"></script>
+    <script defer type="module" src="./js/post.js"></script>
+    <script defer src="./js/script.js"></script>
+
+    <!-- ✅ Bootstrap CSS -->
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
+    />
+
+    <!-- ✅ Favicon -->
+    <link rel="icon" href="./asset/favicon/favicon.ico" />
+    <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="./asset/favicon/apple-touch-icon.png"
+    />
+    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
+
+    <!-- ✅ OG 태그 -->
+    <meta
+      property="og:title"
+      content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
+    />
+    <meta
+      property="og:description"
+      content="우리 아이와 여행을 떠나고 싶다면? PETTY가 도와줄게요 🐾"
+    />
+    <meta
+      property="og:image"
+      content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
+    />
+
+    <!-- ✅ 스타일 -->
+    <link rel="stylesheet" href="common.css" />
+    <link rel="stylesheet" href="write.css" />
+
+    <title>PETTY | 글 작성하기</title>
+  </head>
+
+  <body>
+    <!-- ✅ 네비게이션 -->
+    <div id="navbar"></div>
+
+    <!-- ✅ 상단 배너 -->
+    <section class="container mt-4">
+      <img
+        src="./asset/006.png"
+        class="img-fluid"
+        alt="반려동물 커뮤니티 배너"
+      />
+    </section>
+
+    <!-- ✅ 글 작성 폼 -->
+    <main class="container mt-5 form-section">
+      <div class="card shadow-lg p-4">
+        <div class="card-body text-center">
+          <h2 class="card-title mb-3">게시글 작성</h2>
+          <p class="text-muted">
+            반려동물의 사진, 자랑글, 고민글을 나눠보세요!
+          </p>
+
+          <form id="postForm">
+            <!-- 제목 -->
+            <div class="form-group mb-3">
+              <input
+                type="text"
+                id="title"
+                class="form-control"
+                placeholder="제목을 입력하세요"
+                required
+              />
+            </div>
+
+            <!-- 내용 -->
+            <div class="form-group mb-3">
+              <textarea
+                id="content"
+                class="form-control"
+                rows="4"
+                placeholder="내용을 입력하세요"
+                required
+              ></textarea>
+            </div>
+
+            <!-- 이미지 업로드 -->
+            <div class="form-group mb-3">
+              <label for="image" class="form-label">이미지 업로드</label>
+              <input
+                type="file"
+                id="image"
+                class="form-control"
+                accept="image/*"
+              />
+            </div>
+
+            <!-- 제출 -->
+            <button type="submit" class="btn btn-success w-100">
+              게시글 작성
+            </button>
+          </form>
+        </div>
+      </div>
+    </main>
+
+    <!-- ✅ 푸터 -->
+    <div id="footer"></div>
+
+    <!-- ✅ Bootstrap JS -->
+    <script
+      defer
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"
+    ></script>
+  </body>
+</html>
