--- reco.css
@@ -1,206 +1,0 @@
-.sect-body {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  margin-bottom: 20px;
-  align-items: center;
-  text-align: center;
-  justify-items: center;
-  background-color: rgba(212, 212, 212, 0.685);
-  border-radius: 10px;
-}
-
-.form-wrapper {
-  display: flex;
-  flex-direction: column;
-  width: 50%;
-  padding: 10px;
-  margin-right: 10px;
-}
-.input-group {
-  display: flex;
-  flex-direction: row;
-  width: 100%; /* 부모 요소 기준 가로 정렬 */
-  gap: 10px; /* 입력창 사이 간격 */
-  flex: 1;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-}
-.form-wrapper button {
-  align-items: flex-end;
-  justify-content: end;
-}
-
-.dropdown-menu label {
-  display: block;
-  padding: 10px;
-  cursor: pointer;
-}
-.dropdown-menu label:hover {
-  background-color: #f8f9fa;
-}
-
-.map-container {
-  width: 100%;
-  height: 400px; /* 원하는 높이로 설정 */
-  margin-top: 10px;
-}
-/* Responsive Design */
-@media (max-width: 768px) {
-  .sect-body {
-    display: flex;
-    flex-direction: column;
-  }
-}
-
-@media (max-width: 576px) {
-  .sect-body {
-    display: flex;
-    flex-direction: column;
-  }
-}
-
-h4 {
-  font-family: 'HakgyoansimDunggeunmisoTTF-B';
-}
-
-#login-btn.btn.btn-success {
-  background-color: #a3520f;
-  border: none;
-}
-
-#navbarNav.collapse.navbar-collapse {
-  font-family: 'HakgyoansimDunggeunmisoTTF-B';
-}
-
-.form-wrapper {
-  background-color: white;
-}
-
-.container.mt-4.sect-body {
-  font-family: 'HakgyoansimDunggeunmisoTTF-B';
-  background-color: white;
-  color: white;
-}
-.mb-4 {
-  color: #a3520f;
-}
-.mb-3 {
-  text-align: left;
-  color: black;
-  padding: 10px;
-}
-
-.btn.btn-secondary.dropdown-toggle {
-  background-color: #dddddd;
-  color: rgb(82, 82, 82);
-  border: none;
-}
-.btn.btn-secondary {
-  border: none;
-  background-color: #dddddd;
-  color: rgb(82, 82, 82);
-}
-.btn.btn-primary.mt-3 {
-  background-color: #a3520f;
-  color: white;
-  border: none;
-  font-size: 20px;
-}
-.resultbox {
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-  margin: 20px;
-  padding: 10px;
-  align-items: center;
-  text-align: center;
-  justify-items: center;
-  background-color: rgba(212, 212, 212, 0.24);
-  border-radius: 10px;
-}
-/* map.css */
-.infoWindow {
-  font-weight: bold;
-  display: block;
-}
-.bAddr {
-  padding: 5px;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  white-space: nowrap;
-}
-
-.loading-spinner {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 20px;
-}
-
-.loading-text {
-  font-size: 1rem;
-  color: #ffc107; /* 사용자의 메인 컬러 */
-  font-weight: bold;
-  margin-top: 10px;
-}
-
-/* 전체 카드 */
-.info-card {
-  background: white;
-  padding: 20px;
-  border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  width: 90%;
-  max-width: 800px;
-  margin: 20px auto;
-  text-align: center;
-  font-family: 'HakgyoansimDunggeunmisoTTF-B'
-}
-
-/* 이름 */
-.info-name {
-  font-size: 1.5em;
-  color: #ffc107;
-  margin-bottom: 5px;
-}
-
-/* 주소 */
-.info-address {
-  color: #666;
-  font-size: 0.9em;
-  margin-bottom: 15px;
-  font-family: "Pretendard-Regular";
-}
-
-/* 이미지 */
-.info-image {
-  width: 100%;
-  max-width: 400px;
-  border-radius: 8px;
-  margin: 10px 0;
-}
-
-/* 설명 */
-.info-description {
-  font-size: 0.9em;
-  line-height: 2;
-  color: #444;
-  text-align: center;
-  font-family: "Pretendard-Regular";
-}
-
-/* 운영시간 & 전화번호 */
-.info-hours,
-.info-phone {
-  font-size: 0.9em;
-  color: #333;
-  margin-top: 10px;
-}
-
-.info-phone strong {
-  color: #ffc107;
-}
-
