--- signin.html
@@ -1,141 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <!-- favicon -->
-    <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
-    />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="./asset/favicon/apple-touch-icon.png"
-    />
-    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
-    <meta
-      property="og:title"
-      content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
-    />
-    <meta
-      property="og:description"
-      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 최적의 숙소, 관광지, 음식점을 추천해 줄게요! 🐾"
-    />
-    <meta
-      property="og:image"
-      content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
-    />
-    <!-- <link rel="stylesheet" href="./style.css" /> -->
-    <!-- Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-    <!-- Bootstrap Icons CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
-    />
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
-
-    <link rel="stylesheet" href="common.css" />
-  </head>
-  <body>
-    <!-- nav bar - js -->
-    <div id="navbar"></div>
-
-    <div class="container mt-5">
-      <div class="row justify-content-center">
-        <div class="col-md-6">
-          <div class="card shadow-lg">
-            <div class="card-body">
-              <h3 class="text-center">회원가입</h3>
-              <form>
-                <!-- 이름 -->
-                <div class="mb-3">
-                  <label for="name" class="form-label">이름</label>
-                  <input
-                    type="text"
-                    class="form-control"
-                    id="name"
-                    placeholder="이름을 입력하세요"
-                    required
-                  />
-                </div>
-
-                <!-- 이메일 -->
-                <div class="mb-3">
-                  <label for="email" class="form-label">이메일</label>
-                  <input
-                    type="email"
-                    class="form-control"
-                    id="email"
-                    placeholder="example@email.com"
-                    required
-                  />
-                </div>
-
-                <!-- 비밀번호 -->
-                <div class="mb-3">
-                  <label for="password" class="form-label">비밀번호</label>
-                  <input
-                    type="password"
-                    class="form-control"
-                    id="password"
-                    placeholder="비밀번호 입력"
-                    required
-                    onkeyup="validatePassword()"
-                  />
-                </div>
-
-                <!-- 비밀번호 확인 -->
-                <div class="mb-3">
-                  <label for="confirm-password" class="form-label"
-                    >비밀번호 확인</label
-                  >
-                  <input
-                    type="password"
-                    class="form-control"
-                    id="confirm-password"
-                    placeholder="비밀번호 확인"
-                    required
-                    onkeyup="validatePassword()"
-                  />
-                  <small id="password-message"></small>
-                </div>
-
-                <!-- 회원가입 버튼 -->
-                <div class="d-grid">
-                  <button type="submit" class="btn btn-primary">
-                    가입하기
-                  </button>
-                </div>
-
-                <!-- 로그인 링크 -->
-                <div class="text-center mt-3">
-                  <small
-                    >이미 계정이 있으신가요?
-                    <a href="./login.html">로그인</a></small
-                  >
-                </div>
-              </form>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <!-- nav bar - js -->
-    <div id="footer"></div>
-    <script src="./script.js" defer></script>
-
-    <!-- Bootstrap JS (Optional, for features like dropdowns) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  </body>
-</html>
