--- write.html
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="./script.js" defer></script>
-    <script src="https://cdn.jsdelivr.net/npm/@supabase/supabase-js"></script>
-    <!-- favicon -->
-    <link
-      rel="icon"
-      type="image/png"
-      href="./asset/favicon/favicon-96x96.png"
-      sizes="96x96"
-    />
-    <link rel="icon" type="image/svg+xml" href="./asset/favicon/favicon.svg" />
-    <link rel="shortcut icon" href="./asset/favicon/favicon.ico" />
-    <link
-      rel="apple-touch-icon"
-      sizes="180x180"
-      href="./asset/favicon/apple-touch-icon.png"
-    />
-    <link rel="manifest" href="./asset/favicon/site.webmanifest" />
-    <!-- og tag -->
-    <meta
-      property="og:title"
-      content="PETTY - 반려동물과 함께 떠나는 완벽한 여행"
-    />
-    <meta
-      property="og:description"
-      content="우리 아이(🐶🐱)와 여행을 떠나고 싶다면? PETTY가 최적의 숙소, 관광지, 음식점을 추천해 줄게요! 🐾"
-    />
-    <meta
-      property="og:image"
-      content="https://github.com/user-attachments/assets/42edc0f5-5e02-4709-97c5-3009ed6cdeb2"
-    />
-    <!-- <link rel="stylesheet" href="./style.css" /> -->
-
-    <!-- Bootstrap CSS -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-    />
-    <!-- Bootstrap Icons CSS -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css"
-    />
-
-    <title>PETTY | 반려동물과 함께 떠나는 완벽한 여행</title>
-
-    <link rel="stylesheet" href="common.css" />
-    <link rel="stylesheet" href="write.css" />
-  </head>
-  <!-- ====== body ======= -->
-  <body>
-    <!-- nav bar - js -->
-    <div id="navbar"></div>
-
-    <section class="container mt-4">
-      <img src="./asset/006.png" class="img-fluid" />
-    </section>
-
-    <section class="container mt-5 form-section">
-      <div class="card shadow-lg p-4">
-        <div class="card-body text-center">
-          <h2 class="card-title mb-3">게시글 작성</h2>
-          <p class="text-muted">
-            반려동물의 사진, 자랑글, 고민글 등 다양한 이야기를 나눠보세요
-          </p>
-
-          <form id="postForm">
-            <!-- 제목 입력 -->
-            <div class="form-group mb-3">
-              <input
-                type="text"
-                id="title"
-                class="form-control"
-                placeholder="제목을 입력하세요"
-                required
-              />
-            </div>
-
-            <!-- 내용 입력 -->
-            <div class="form-group mb-3">
-              <textarea
-                id="content"
-                class="form-control"
-                rows="4"
-                placeholder="내용을 입력하세요"
-                required
-              ></textarea>
-            </div>
-
-            <!-- 이미지 업로드 -->
-            <div class="form-group mb-3">
-              <label for="image" class="form-label">이미지 업로드</label>
-              <input type="file" id="image" class="form-control" />
-            </div>
-
-            <!-- 제출 버튼 -->
-            <button type="submit" class="btn btn-success w-100">
-              게시글 작성
-            </button>
-          </form>
-        </div>
-      </div>
-    </section>
-
-    <!-- nav bar - js -->
-    <div id="footer"></div>
-    <script src="./script.js" defer></script>
-    <script src="./main.js"></script>
-    <script src="./post.js"></script>
-    <!-- Bootstrap JS (Optional, for features like dropdowns) -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
-  </body>
-</html>
